Refactor VisitWizard to use PageBase
